refactor: single purpose utils
